# Optimize string (de)serialization on .NET Core.